Refactor tournament logic into separate modules
